requirements.txt
essentia-tensorflow==2.1b6.dev1389
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
hdbscan>=0.8.33
plotly>=5.14.0
sentence-transformers>=2.2.2
langdetect>=1.0.9
tqdm>=4.65.0
